third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/fetch-request-xhr-sync.https-expected.txt

This is a testharness.js-based test.
FAIL Verify SyncXHR is interceptedassert_equals: HTTP response status code for intercepted request expected 200 but got 404
Harness: the test ran to completion.
